+++ src/general/_index.js
@@ -1,4 +1,5 @@
 import './breakpoints/breakpoints.init';
 import './helpers/date_and_time/date_and_time.init';
+import './icons/icons.init';
 import './helpers/spacing/spacing.init';
 import './typography/typography.init';